Break up the legacyBfacf.h monolith header file
