+++ src/clkTopology.h
@@ -1,6 +1,8 @@
 #pragma once
 
-#include "legacyBfacf.h"
+#include "clkTypes.h"
+#include "clkConstants.h"
+#include "clkTables.h"
 
 // Topology detection
 EdgePtr is_tight_clasp(CubicLatticeKnotPtr clkp, EdgePtr ep);
