--- src/legacyBfacf.cpp
@@ -1,10 +1,0 @@
-#include "legacyBfacf.h"
-
-using namespace std;
-
-extern void set_sRand_seed_to_clocktime();
-extern int rand_integer(int, int);
-extern double rand_double(double low, double high);
-extern double rand_uniform();
-extern void sRandSimple(int seed);
-
